Fix stabilizers for groups-with-memory
Computing stabilizers in groups-with-memory could result in
unexpected errors where we ended up trying to compute e.g. characteristic
polynomials of matrices-with-memory. The problem was this code in
`GENSS_FindVectorsWithShortOrbit`:

    if IsObjWithMemory(l[1]) then
        ForgetMemory(l);
    fi;

It can happen that the first element of `l` was a matrix without memory,
but later elements *did* have a memory. In that case, we failed to call
`ForgetMemory` and thus the problem.

The fix is to simply always invoke `ForgetMemory`.